Detect whether Opam is installed

+++ src/lib/opam.ml
@@ -1,20 +1,23 @@
 open! Import
 open Bos_setup
 
-let install ~yes:_ () =
-  (* FIXME: implement this in OCaml instead of running the opam script and adapt it to our ideas. In particular, take [yes] into account instead of ignoring it :P *)
-  let install_sh =
-    "bash -c \"sh <(curl -fsSL \
-     https://raw.githubusercontent.com/ocaml/opam/master/shell/install.sh)\""
-  in
-  let exit = Sys.command install_sh in
-  if exit = 0 then Ok ()
-  else Error (`Msg "Something went wrong trying to install opam.")
+let is_installed () = OS.Cmd.exists (Cmd.v "opam")
 
-let is_installed () =
-  (*FIXME*)
-  false
-(* FIXME: implement this in OCaml instead of running the opam script *)
+let install ~yes:_ () =
+  let open Result.Direct in
+  let* installed = is_installed () in
+  if installed then Error (`Msg "Already installed")
+  else
+    (* FIXME: implement this in OCaml instead of running the opam script and
+       adapt it to our ideas. In particular, take [yes] into account instead of
+       ignoring it :P *)
+    let install_sh =
+      "bash -c \"sh <(curl -fsSL \
+       https://raw.githubusercontent.com/ocaml/opam/master/shell/install.sh)\""
+    in
+    let exit = Sys.command install_sh in
+    if exit = 0 then Ok ()
+    else Error (`Msg "Something went wrong trying to install opam.")
 
 let add_yes ~yes cmd = if yes then Cmd.(cmd % "--yes") else cmd
